requirements.txt
tqdm~=4.64.0
Pillow~=9.2.0
setuptools~=58.1.0
json5~=0.9.10